requirements.txt
Django==2.1.2
django-rest-framework==0.1.0
djangorestframework==3.8.2
psycopg2-binary==2.7.5
psycopg2==2.7.5
django-filter==2.0.0
djangorestframework-queryfields==1.0.0